blog/urls.py
# blog/urls.py

from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import CategoryViewSet, PostViewSet, CommentViewSet
from .auth.views import AuthViewSet

router = DefaultRouter()

router.register(r'categories', CategoryViewSet, basename='category')
router.register(r'posts', PostViewSet, basename='post')
router.register(r'comments', CommentViewSet, basename='comment')
router.register(r'auth', AuthViewSet, basename='auth')

urlpatterns = [
    path('', include(router.urls)),
    path('posts/<slug:post_slug>/comments/', CommentViewSet.as_view({
        'get': 'list',
        'post': 'create',
    }), name='post-comments'),
    path('posts/<slug:post_slug>/comments/<int:pk>/', CommentViewSet.as_view({
        'get': 'retrieve',
        'put': 'update',
        'patch': 'partial_update',
        'delete': 'destroy', 
    }), name='post-comment-detail'),
]

