requirements.txt
django==5.2.1
mysqlclient
python-decouple
djangorestframework
unidecode==1.3.6
